chore(planes): refactorizacion comandos y queries
