+++ src/planes/application/commands/asociar_entrenamiento.py
@@ -1,25 +1,19 @@
-import uuid
 import traceback
+import uuid
 from dataclasses import dataclass, field
 
 from sqlalchemy.exc import IntegrityError
 
 from planes.application.commands.base import PlanCommandBaseHandler
+from planes.application.dtos import PlanEntrenamientoDTO
+from planes.application.exceptions import BadRequestError, UnprocessableEntityError
 from planes.application.mappers import PlanEntrenamientoDTOEntityMapper
+from planes.domain.entities import PlanEntrenamiento
+from planes.infrastructure.uwo import UnitOfWorkASQLAlchemyFactory
 from seedwork.application.commands import Command, execute_command
 from seedwork.domain.exceptions import BusinessRuleException
 from seedwork.infrastructure.uow import UnitOfWorkPort
 from seedwork.presentation.exceptions import APIError
-from planes.application.dtos import EntrenamientoDTO, PlanEntrenamientoDTO
-from planes.application.exceptions import (
-    UnprocessableEntityError,
-    BadRequestError,
-)
-from planes.domain.entities import (
-    Entrenamiento,
-    PlanEntrenamiento,
-)
-from planes.infrastructure.uwo import UnitOfWorkASQLAlchemyFactory
 
 
 @dataclass
