+++ src/planes/application/exceptions.py
@@ -1,4 +1,5 @@
 from http import HTTPStatus
+
 from seedwork.presentation.exceptions import APIError
 
 
@@ -11,10 +12,14 @@ class BadRequestError(APIError):
     def __init__(self, message=HTTPStatus.BAD_REQUEST.phrase, code=""):
         super().__init__(HTTPStatus.BAD_REQUEST.value, message, code)
 
+
 class UserNotFoundError(APIError):
     def __init__(self, message=HTTPStatus.NOT_FOUND.phrase):
         super().__init__(HTTPStatus.NOT_FOUND.value, message, "user.not_found")
 
+
 class UserNotAsociatedError(APIError):
     def __init__(self, message=HTTPStatus.UNPROCESSABLE_ENTITY.phrase):
-        super().__init__(HTTPStatus.UNPROCESSABLE_ENTITY.value, message, "user.not_asociated_to_plan")
+        super().__init__(
+            HTTPStatus.UNPROCESSABLE_ENTITY.value, message, "user.not_asociated_to_plan"
+        )
