+++ src/planes/infrastructure/auth.py
@@ -1,4 +1,3 @@
 from flask_jwt_extended import JWTManager
 
-
 jwt = JWTManager()
