+++ src/planes/infrastructure/db.py
@@ -1,5 +1,4 @@
 from flask import Flask
 from flask_sqlalchemy import SQLAlchemy
 
-
 db = SQLAlchemy()
